requirements.txt
google-api-python-client>=2.129.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pdfplumber>=0.11.4
pillow>=10.4.0
pdfminer.six>=20231228
openai>=1.40.0
python-dotenv>=1.0.1